세정장고231018/10-02-django-rest-framework/requirements.txt
asgiref==3.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
idna==3.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.31.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toposort==1.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
